packers_genai_bot/requirements.txt
streamlit
openai
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
pandas
xlrd
fpdf
mysql.connector
